refactor(fit): break persistence-manager import cycle
Move fit metadata and registry models into schema.dart so persistence depends only on storage schemas instead of the Riverpod manager layer.
