+++ lib/storage/fit/manager.dart
@@ -11,39 +11,11 @@ import "package:eve_fit_assistant/utils/riverpod.dart";
 import "package:fast_immutable_collections/fast_immutable_collections.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:fpdart/fpdart.dart";
-import "package:freezed_annotation/freezed_annotation.dart";
 import "package:path/path.dart" as p;
 import "package:riverpod_annotation/riverpod_annotation.dart";
 import "package:uuid/uuid.dart";
 
-part "manager.freezed.dart";
 part "manager.g.dart";
-
-@freezed
-abstract class FitMetadata with _$FitMetadata {
-  const factory FitMetadata({
-    required String fitId,
-    required int shipTypeId,
-    required String name,
-
-    /// DateTime.millisecondsSinceEpoch
-    required int lastModified,
-
-    required String description,
-    required String bundleId,
-  }) = _FitMetadata;
-
-  factory FitMetadata.fromJson(Map<String, dynamic> json) => _$FitMetadataFromJson(json);
-}
-
-@freezed
-abstract class FitRegistry with _$FitRegistry {
-  const factory FitRegistry({
-    @JsonKey(defaultValue: IMap.empty) required IMap<String, FitMetadata> fits,
-  }) = _FitRegistry;
-
-  factory FitRegistry.fromJson(Map<String, dynamic> json) => _$FitRegistryFromJson(json);
-}
 
 /// Fit storage is always under global control,
 /// So there's no need to maintain a global singleton outside of the Ref tree.
